+++ crates/kreuzberg/src/extractors/excel.rs
@@ -102,7 +102,7 @@ impl ExcelExtractor {
         let mut doc = Self::build_internal_document(workbook);
 
         let sheet_names: Vec<String> = workbook.sheets.iter().map(|s| s.name.clone()).collect();
-        let sheet_count = workbook.sheets.len();
+        let sheet_count = workbook.sheets.len() as u32;
         let excel_metadata = ExcelMetadata {
             sheet_count: Some(sheet_count),
             sheet_names: Some(sheet_names),
